break out hybrid_dispatch_options.py and make Dispatch class instantiation based on RE type more generalized
